+++ connectors/bindings.amqp/test/queues.mock.js
@@ -4,4 +4,8 @@ import { generate } from 'randomstring'
 
 export const name = mock.fn(() => generate())
 
+export const tasks = mock.fn(() => generate())
+
 export const instances = mock.fn(() => generate())
+
+export const scoped = mock.fn((name) => name)