+++ extensions/exposition/features/steps/components/sequences/operations/failing.js
@@ -1,0 +1,5 @@
+export async function* computation() {
+  yield 'first'
+
+  throw new Error('The sequence is broken')
+}